Ports Baron Wanted (#425)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

<!-- NOTE: This format is NOT REQUIRED for things like runtime fixes,
code fixes and optimizations. In those instances you should try to give
a description of what is being fixed or optimized but are not required
to fill out the complete changelog. -->
<!-- Adjusting things like armor or weapon values for balance, while
they may be 'fixes' in their own way, are not considered code fixes as
described above and require the use of the Pull Request format below.-->


## About The Pull Request

Two ports - one from Paradise Baron requested that makes it so, upon
saying a staffs name (exactly, case sensative) in staff chat, the other
staff will get a 'ping' noise that plays.

The other is from Coyote Bayou. It's their older version of their
'profile picture' code. The reason why this isn't the current up-to-date
one is, upon trying to port it exactly as they have it, the popup to
insert a link refused to show and kept acting as if I was removing a
picture despite never having uploaded one.

So - it'll do for now. Enjoy having it at all, I guess.

Also, found a version of Bubber's age verification idea that works on
our codebase. Ported it over. This allows staff to set someone as
age-checked if they ID verify them in the discord.

## Why It's Good For The Game

Requested ports, both are decent features. Latter one could use some
development if anyone knows how to fix it to modern up-to-date CB's
profile pictures.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ X ] You tested this on a local server.
- [ X ] This code did not runtime during testing.
- [ X ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->


## Changelog
<!-- This is mostly optional for small Pull Requests, but if you value
being credited for your work in-game be sure to fill it out. To opt-out,
remove everything below including the start and end :cl: brackets. -->

<!-- If your Pull Request includes a minor single line variable edit,
probably do not fill out this changelog. -->
<!-- However, if your pull request includes massive or immediately
noticeable changes, briefly describe those changes in some way in this
changelog. -->

:cl:
adds: Paradise station admin ping system
adds: CB's profile picture (old) system
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->