Fixes Power Plant Access (#476)
## About The Pull Request
#461 changed power plant access from the power plant tag to the town
tag, which let more town roles in but locked the followers out. This
works better - restores the power plant tag, but gives it to the town
roles that were missing it.

Also removes a power plant door that was just stuck in a wall. : )
## Why It's Good For The Game
Now everyone in Town and Follower factions can access the plant. Problem
solved.
## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
## Changelog
:cl:
fix: Fixes Bighorn Power Plant access.
/:cl: