Fixes Jukeboxes (#534)
## About The Pull Request
Jukeboxes are finally back with 33 different songs. Full credit to
MrMatrixMan for this.

They already have some decent placement on the map, but added one to the
Bighorn Bar. I would have added one to the NCR base but that map file is
being changed by #531.

## Why It's Good For The Game
Now people can listen to music!

There is some work to do in future with faction-specific
jukeboxes/tapes, but this is a good basic level to add it at.
## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
## Changelog
:cl:
fix: Fixed jukeboxes. Enjoy the music!
/:cl: