Warrens remake! (#531)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

<!-- NOTE: This format is NOT REQUIRED for things like runtime fixes,
code fixes and optimizations. In those instances you should try to give
a description of what is being fixed or optimized but are not required
to fill out the complete changelog. -->
<!-- Adjusting things like armor or weapon values for balance, while
they may be 'fixes' in their own way, are not considered code fixes as
described above and require the use of the Pull Request format below.-->


## About The Pull Request

This Pull request remakes the East side of Warrens. 
Mainly the NCR base (moved up north)
and the area where the NCR base was
Moves the raiders to the sewers

Balance wise, it moves the High weapon spawner from the Renegades, to
the new Raider captured NCR food processing plant, adds a RCA spawner, a
high-mid spawner and some misc other minor loots.

it also adds a lot more flavour to the warren sector, showing that the
NCR has a presence, outside their outpost
( the NCR housing right outside their gate / their cannery, now taken
over by raiders / the small farm above the cannery

the raiders now get to live more.... "raidery?", with their underground
city!


![image](https://github.com/Stutternov/f13tbd/assets/79304582/86df86dd-ad78-4ecb-b9a7-405279c2fd8e)


![image](https://github.com/Stutternov/f13tbd/assets/79304582/8af2f48e-0cf7-4ca5-9097-11a9587c567c)



![StrongDMM-2024-01-17 15 14
23](https://github.com/Stutternov/f13tbd/assets/79304582/d43b8478-9d03-4bab-8dea-bc9df716244d)

![StrongDMM-2024-01-17 15 14
25](https://github.com/Stutternov/f13tbd/assets/79304582/87e41087-c238-46cb-9be5-7d8b17af7b75)


![image](https://github.com/Stutternov/f13tbd/assets/79304582/78b43ea1-fcc5-4a51-8343-1fb177093350)


![image](https://github.com/Stutternov/f13tbd/assets/79304582/2e16e071-3200-440c-b9ec-96b5966be283)


![image](https://github.com/f13babylon/f13babylon/assets/79304582/7c6e1951-24f2-4cd7-8ca5-494f03d17fe6)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
The current NCR base is GARGANTUAN, and doesn't do much with that space.
it's annoying to attack, can be painful to roleplay in, and in general,
is just sub-par.

The new one is much more compact, with a way more "realistic" design as
to how the NCR may operate such a outpost.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->


## Changelog
<!-- This is mostly optional for small Pull Requests, but if you value
being credited for your work in-game be sure to fill it out. To opt-out,
remove everything below including the start and end :cl: brackets. -->

<!-- If your Pull Request includes a minor single line variable edit,
probably do not fill out this changelog. -->
<!-- However, if your pull request includes massive or immediately
noticeable changes, briefly describe those changes in some way in this
changelog. -->

:cl:
tweak: Moves the NCR base up North
tweak: Moves the raiders into the sewers
:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->