Gun Sound Overhaul (#536)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

<!-- NOTE: This format is NOT REQUIRED for things like runtime fixes,
code fixes and optimizations. In those instances you should try to give
a description of what is being fixed or optimized but are not required
to fill out the complete changelog. -->
<!-- Adjusting things like armor or weapon values for balance, while
they may be 'fixes' in their own way, are not considered code fixes as
described above and require the use of the Pull Request format below.-->


## About The Pull Request

Basically modularizes parts of the sounds of guns. Before it would call
a specific gun sound defined in sound.dm and randomly choose between 4
different generic gun sounds for empty mags, full mags, and gun cocking.
Now instead - it's a set sound.

Mags now make a sound when loaded, when unloaded, and if a round enters
the chamber when loaded or is removed from the chamber after unloading
there is also a new sound for this.

These sounds can be changed per-gun path if desired or even var-edited
in game now to provide unique gun sounds.

These include:
- Pistol, SMG, heavy pistol, light rifle, rifle, LMG, and Minigun reload
noises.
- Pistol, SMG, heavy pistol, light rifle, rifle, LMG, and Minigun
unloading noises.
- Pistol, rifle, lmg, bolt action, and shotgun slide sounds. (Shotgun
slide is actually a break-action sound.)
- New bolt-cycle sound for the pump-action on rifles. No more pump
action shotgun sound for them.
- New shell insert noises for both pump action shotguns and automatic
internal loading shotgun types.

For **now**, empty magazines sound identical to full ones when inserted.
The var exists to replace this sound with a unique one if wanted, but I
couldn't find any good sound files that meshed well with the loaded mag
sounds.

Most of the sounds here were either provided by Poots or ported from
Eris.

## Why It's Good For The Game

Redoes some of the soundscape. This is a good first step, as next would
be redoing some sprites for ammo boxes, ammo casings, and likely more
sprite changes. Good step in the right direction for these changes to
make the game sound more immersive and allow modularity for staff when
hosting events.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ X ] You tested this on a local server.
- [ X ] This code did not runtime during testing.
- [ X ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->


## Changelog
<!-- This is mostly optional for small Pull Requests, but if you value
being credited for your work in-game be sure to fill it out. To opt-out,
remove everything below including the start and end :cl: brackets. -->

<!-- If your Pull Request includes a minor single line variable edit,
probably do not fill out this changelog. -->
<!-- However, if your pull request includes massive or immediately
noticeable changes, briefly describe those changes in some way in this
changelog. -->

:cl:
add: New gun sound varaibles to allow the changing of sound files with
ease.
add: New gun reload noises, unload noises, gun cocking noises, bolt
cycling noises, and shell inserting noises.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Signed-off-by: Stutternov <47883419+Stutternov@users.noreply.github.com>
Co-authored-by: Tom <8881105+tf-4@users.noreply.github.com>